moved premium.component in web from base settings to a new settings folder inside vault
