+++ apps/web/src/app/billing/settings/subscription-routing.module.ts
@@ -4,7 +4,7 @@ import { RouterModule, Routes } from "@angular/router";
 import { BillingHistoryViewComponent } from "../../billing/settings/billing-history-view.component";
 import { PaymentMethodComponent } from "../../billing/settings/payment-method.component";
 import { UserSubscriptionComponent } from "../../billing/settings/user-subscription.component";
-import { PremiumComponent } from "../../settings/premium.component";
+import { PremiumComponent } from "../../vault/settings/premium.component";
 
 import { SubscriptionComponent } from "./subscription.component";
